finpudsning

+++ Project/ProjectManager/src/projectManager/Activity.java
@@ -8,10 +8,10 @@ public class Activity {
 	private String description;
 	private Date startDate;
 	private Date endDate;
-	private int timeBudget;
+	private double timeBudget;
 	private List<User> users = new ArrayList<User>();
 	
-	public Activity(String name, String description, Date startDate, Date endDate, int timeBudget) {
+	public Activity(String name, String description, Date startDate, Date endDate, double timeBudget) {
 		this.name = name;
 		this.description = description;
 		this.startDate = startDate;
@@ -34,14 +34,14 @@ public class Activity {
 	public Date getEndDate() {
 		return endDate;
 	}
-	public int getRemainingHours() {
-		int hours = 0;
+	public double getRemainingHours() {
+		double hours = 0;
 		for(User user : users){
 			hours = hours + user.getRegisteredTime(this);
 		}
 		return timeBudget - hours;
 	}
-	public int getTimeBudget() {
+	public double getTimeBudget() {
 		return timeBudget;
 	}
